personalizado e implementado com desafio

+++ modulo2/apis-labenusers/src/components/TelaListaUsuarios.jsx
@@ -5,16 +5,26 @@ import styled from "styled-components";
 
 
 const CardUsuario = styled.div`
-       color: green;
+      color: green;
        padding: 10px;
        width: 300px;
        display:flex;
        justify-content: space-between;
+
+       button {
+        width: 8%;
+        display:flex;
+        justify-content: center;
+        background-color:red;
+       }
 `
+
+
+
 export default class TelaListaUsuarios extends React.Component {
-    state= {
-      usuarios: []
-    }
+  state = {
+    usuarios: []
+  }
 
   componentDidMount() {
     this.pegarUsuarios()
@@ -30,7 +40,7 @@ export default class TelaListaUsuarios extends React.Component {
       }
     })
       .then((res) => {
-        this.setState({usuarios:res.data})
+        this.setState({ usuarios: res.data })
       })
 
       .catch((err) => {
@@ -41,30 +51,34 @@ export default class TelaListaUsuarios extends React.Component {
 
 
   deletarUsuario = (id) => {
-    const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`
-    
-    axios.delete(url, {
-      headers: {
-        Authorization: "Gustavo-Monteiro"
-      }
-    }).then((res)=>{
-      alert
-      ('Usuário(a) deletado(a) com sucesso!')
-      this.pegarUsuarios()
-    })
-    .catch((err)=>{
-      alert('Ocorreu um erro, tente novamente')
-    })
-  }
+    if (window.confirm('Tem certeza que deseja deletar?')) {
+      const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`
 
+      axios.delete(url, {
+        headers: {
+          Authorization: "Gustavo-Monteiro"
+        }
+      }).then((res) => {
+        alert
+          ('Usuário(a) deletado(a) com sucesso!')
+        this.pegarUsuarios()
+      })
+        .catch((err) => {
+          alert('Ocorreu um erro, tente novamente')
+        })
+    }
+    else {
+      alert('operação cancelada')
+    }
+  }
 
   render() {
     console.log(this.state.usuarios)
-    const listaUsuarios = this.state.usuarios.map((user)=> {
-        return <CardUsuario key={user.id}>
-          {user.name}
-          <button onClick={()=> this.deletarUsuario(user.id)}>x</button>
-          </CardUsuario>
+    const listaUsuarios = this.state.usuarios.map((user) => {
+      return <CardUsuario key={user.id}>
+        {user.name}
+        <button onClick={() => this.deletarUsuario(user.id)}>x</button>
+      </CardUsuario>
     })
